NonNull어노테이션을 NotNull 어노테이션으로 변경
@Valid로 null 데이터를 막기 위해서 변경

참고: #25
